#32 [REFACTOR] 카테고리 라벨 단일 소스로 통합 (constants/explore.ts)
